feat: Allow purging package list cache.
When you install packages, if you need to update the package
manager cache, a lot of data may be pulled down to your
system (in the 100s of MB). This is quite cumbersome, especially
for container images.

Setting this role variable `package_cache_purge` to `true` will
ensure you don't leave all that useless data behind.